Refactoring calendar code and add tests for calendar
