+++ src/assets/ts/lib/calendar/calendarConstruction.ts
@@ -1,9 +1,9 @@
 import {Moment} from 'moment';
 
-export interface CalendarConstruction {
-    date: Moment;
-    monthBox: HTMLSpanElement;
-    yearBox: HTMLSpanElement;
-    daysTable: HTMLTableSectionElement;
-    mapDataToCalendar():void;
-}
+// export interface CalendarConstruction {
+//     date: Moment;
+//     monthBox: HTMLSpanElement;
+//     yearBox: HTMLSpanElement;
+//     daysTable: HTMLTableSectionElement;
+//     mapDataToCalendar():void;
+// }
