+++ src/index.html
@@ -27,32 +27,6 @@
         <div class="box">
             <h2>Calendar</h2>
             <section id="calendar">
-                <nav>
-                    <span id="prev">
-                        <i class="fas fa-chevron-left"></i>
-                    </span>
-                    <span id="month">March</span>
-                    <span id="year">2021</span>
-                    <span id="next">
-                        <i class="fas fa-chevron-right"></i>
-                    </span>
-                </nav>
-                <table>
-                    <thead>
-                        <tr>
-                            <th>Mon</th>
-                            <th>Tue</th>
-                            <th>Wed</th>
-                            <th>Thu</th>
-                            <th>Fri</th>
-                            <th>Sat</th>
-                            <th>Sun</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-        
-                    </tbody>
-                </table>
             </section>   
         </div>
     </div>
